requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Werkzeug==3.0.1
WTForms==3.1.1
python-dotenv==1.0.0
geopy==2.4.1
numpy<2
opencv-python==4.10.0.84
requests==2.31.0
